fix(messaging/feishu): resolve card header branch on first turn
Lazy-resolve git branch from workDir when lastBranch is empty,
extract GitBranchOf as shared utility to eliminate code duplication.

Closes #314
